Moved 'Query References' component to its own file.
